print numbers using putchar

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -9,9 +9,9 @@
 
 void print_number(int n)
 {
-unsigned int m;
+long m;
 int c;
-unsigned int num;
+long num;
 
 num = n;
